requirements.txt
# Database
psycopg2-binary>=2.9.11
sqlalchemy>=2.0.0
alembic>=1.13.0
pgvector>=0.2.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.17.2
scrapy==2.11.0
lxml==5.1.0

# HTML Parsing
html5lib==1.1

# Email
jinja2==3.1.3

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.1

# Logging
colorlog==6.8.2

# Data Validation
pydantic>=2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0  # FastAPI 테스트용
requests>=2.31.0  # 통합 테스트용

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Google Generative AI (Gemini API)
google-generativeai>=0.3.0

# Monitoring Web Dashboard
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
docker>=7.0.0
